+++ front-end/src/components/menu-lateral/contexts/DrawerContext.tsx
@@ -1,0 +1,43 @@
+import { createContext, useCallback, useContext, useState } from 'react';
+
+interface IDrawerOption {
+    icon: string;
+    path: string;
+    label: string;
+}
+
+interface IDrawerContextData {
+    isDrawerOpen: boolean;
+    toggleDrawerOpen: () => void;
+    drawerOptions: IDrawerOption[];
+    setDrawerOptions: (newDrawerOptions: IDrawerOption[]) => void;
+
+}
+
+const DrawerContext = createContext({} as IDrawerContextData);
+
+export const useDrawerContext = () => {
+    return useContext(DrawerContext);
+};
+
+interface IDrawerProviderProps {
+    children: React.ReactNode
+}
+export const DrawerProvider: React.FC<IDrawerProviderProps> = ({ children }) => {
+    const [drawerOptions, setDrawerOptions] = useState<IDrawerOption[]>([]);
+    const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+
+    const toggleDrawerOpen = useCallback(() => {
+        setIsDrawerOpen(oldDrawerOpen => !oldDrawerOpen);
+      }, []);
+
+      const handleSetDrawerOptions = useCallback((newDrawerOptions: IDrawerOption[]) => {
+        setDrawerOptions(newDrawerOptions);
+      }, []);
+
+    return (
+        <DrawerContext.Provider value={{ isDrawerOpen, drawerOptions, toggleDrawerOpen, setDrawerOptions: handleSetDrawerOptions }}>
+            {children}
+        </DrawerContext.Provider>
+    );
+};
