+++ front-end/src/pages/gestaoDeAcesso/index.tsx
@@ -10,7 +10,6 @@ import Input from '../../components/Input';
 import { fetchGestores } from '../../services/gestores';
 import { cadastrarGestor } from '../../controller/cadastrarGestor';
 import { atribuirGestorProjeto } from '../../controller/linkGestorProjeto';
-import MenuLateral from '../menu-lateral/MenuLateral';
 
 export default function GestaoDeAcesso() {
     const [gestoresList, setGestoresList] = useState<ResponseAdminApi[]>([]);
@@ -111,7 +110,6 @@ export default function GestaoDeAcesso() {
 
     return (
         <>
-            <MenuLateral />
             <Stack spacing={3} margin="2% auto" marginLeft="25%">
