+++ front-end/src/routes/routes.tsx
@@ -1,4 +1,5 @@
 import { BrowserRouter } from 'react-router-dom'
+import MenuLateral from '../components/menu-lateral/MenuLateral'
 import { AppRoutes } from './app.routes'
 
 
@@ -7,6 +8,7 @@ export function Routes() {
 
         <BrowserRouter>
             <AppRoutes />
+            <MenuLateral />
         </BrowserRouter>
     )
 }
